Extract propKeys() into separate file
